requirements.txt
asgiref==3.5.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-cors-headers==3.12.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.6.0
gunicorn==20.1.0
idna==3.3
phonenumbers==8.12.49
Pillow==9.1.1
platformdirs==2.5.2
psycopg-binary==3.0.15
psycopg2==2.9.3
PyJWT==2.4.0
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.2.0
